third_party/WebKit/LayoutTests/platform/win/external/wpt/webmessaging/event.ports.sub-expected.txt

This is a testharness.js-based test.
PASS Test Description: event.ports returns the MessagePort array sent with the message. 
PASS MessageChannel is supported. 
Harness: the test ran to completion.
